move default/base model to defaults folder
