+++ src/ewatercycle/plugins/marrmot/marrmot.py
@@ -13,7 +13,7 @@ from cftime import num2date
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.models.abstract import DefaultModel
+from ewatercycle.defaults.model import DefaultModel
 from ewatercycle.plugins.marrmot.forcing import MarrmotForcing
 from ewatercycle.util import get_time, to_absolute_path
